PL-75: Simplify audit field resolvers
